LayoutTests/imported/w3c/web-platform-tests/css/css-values/attr-cycle-expected.txt


PASS CSS Values and Units Test: attr
PASS CSS Values and Units Test: attr 1
PASS CSS Values and Units Test: attr 2
PASS CSS Values and Units Test: attr 3
PASS CSS Values and Units Test: attr 4
PASS CSS Values and Units Test: attr 5
PASS CSS Values and Units Test: attr 6
PASS CSS Values and Units Test: attr 7
PASS CSS Values and Units Test: attr 8
PASS CSS Values and Units Test: attr 9
PASS CSS Values and Units Test: attr 10
FAIL CSS Values and Units Test: attr 11 assert_equals: Value 'attr(data-foo type(<custom-ident>), 3px)', where 'data-foo=var(--y)' should be valid for the property '--x'. expected "3px" but got ""
FAIL CSS Values and Units Test: attr 12 assert_equals: Value 'attr(data-foo type(<length>), 11px)', where 'data-foo=attr(data-foo type(<length>), 3px)' should be valid for the property '--x'. expected "11px" but got ""
FAIL CSS Values and Units Test: attr 13 assert_equals: Value 'attr(data-foo type(<length>), 11px)', where 'data-foo=attr(data-foo type(<length>))' should be valid for the property '--x'. expected "11px" but got ""
FAIL CSS Values and Units Test: attr 14 assert_equals: Value 'attr(data-foo type(*), 11px)', where 'data-foo=attr(data-foo type(*))' should be valid for the property '--y'. expected "11px" but got ""
FAIL CSS Values and Units Test: attr 15 assert_equals: Value 'attr(data-foo type(<length>), attr(data-bar type(<length>)))', where 'data-foo=attr(data-foo type(*))' should be valid for the property '--x'. expected "11px" but got ""
FAIL CSS Values and Units Test: attr 16 assert_equals: Value 'attr(data-foo type(*), attr(data-bar))', where 'data-foo=attr(data-foo type(*))' should be valid for the property '--y'. expected "\"abc\"" but got ""
FAIL CSS Values and Units Test: attr 17 assert_equals: Value 'attr(data-foo type(*), 1px)', where 'data-foo=attr(data-bar type(*), 2px)' should be valid for the property '--y'. expected "1px" but got ""
PASS CSS Values and Units Test: attr 18
PASS CSS Values and Units Test: attr 19
FAIL CSS Values and Units Test: attr 20 assert_equals: Value 'attr(data-foo type(*), var(--y))', where 'data-foo=3' should be valid for the property '--y'. expected "3" but got ""
FAIL CSS Values and Units Test: attr 21 assert_equals: Value 'attr(data-foo type(*), attr(data-foo))', where 'data-foo=3' should be valid for the property '--y'. expected "3" but got ""
FAIL CSS Values and Units Test: attr 22 assert_equals: Value 'attr(data-foo type(*), var(--x))', where 'data-foo=3' should be valid for the property '--y'. expected "3" but got ""
FAIL CSS Values and Units Test: attr 23 assert_equals: Value 'attr(data-foo type(*), attr(data-bar type(*)))', where 'data-foo=3' should be valid for the property '--y'. expected "3" but got ""
PASS CSS Values and Units Test: attr 24
PASS CSS Values and Units Test: attr 25
FAIL CSS Values and Units Test: attr 26 assert_equals: Value 'attr(data-foo type(*))', where 'data-foo=attr(data-bar, 11) var(--x, 3)' should be valid for the property '--y'. expected "\"var(--x)\" \"attr(data-bar, 11) var(--x, 3)\"" but got ""
FAIL CSS Values and Units Test: attr 27 assert_equals: Value 'attr(data-foo type(*), abc)', where 'data-foo=attr(data-foo' should be valid for the property '--x'. expected "abc" but got ""
